Uses configure.ac generated from autoscan
We know from the file list at􏰀that config.h.in has been created, so we know that autoreconf has executed autoheader. We also see the new install-sh script at􏰁that was created when we executed automake in autogen.sh. Anything provided or generated by the Autotools should be copied into the archive directory so that it can be shipped with release tarballs. Therefore, we’ll add cp commands for these two files to the $(distdir) target in the top-level Makefile.in template. Note that we don’t need to copy the autogen.sh script because it’s purely a maintainer tool—users should never need to execute it from a tarball distribution.